+++ src/Settings.ts
@@ -672,6 +672,7 @@ export const SETTINGS_SOURCEPATH_PREFIX = "@Code-Styler-Settings:";
 export const LOCAL_PREFIX = "@/";
 export const REFERENCE_CODEBLOCK = "reference";
 export const EXTERNAL_REFERENCE_PATH = "/plugins/code-styler/reference-files/";
+export const EXTERNAL_REFERENCE_CACHE = EXTERNAL_REFERENCE_PATH + "cache.json";
 export const EXTERNAL_REFERENCE_INFO_SUFFIX = "-info.json";
 export const GIT_ICONS: { [key: string]: string } = {
 	"branch": "&#xe0a0;",