+++ src/main.ts
@@ -1,6 +1,6 @@
 import { Plugin, MarkdownView, WorkspaceLeaf } from "obsidian";
 
-import { convertSettings, DEFAULT_SETTINGS, LANGUAGES, CodeStylerSettings, REFERENCE_CODEBLOCK, EXTERNAL_REFERENCE_PATH } from "./Settings";
+import { convertSettings, DEFAULT_SETTINGS, LANGUAGES, CodeStylerSettings, REFERENCE_CODEBLOCK, EXTERNAL_REFERENCE_PATH, EXTERNAL_REFERENCE_CACHE } from "./Settings";
 import { SettingsTab } from "./SettingsTab";
 import { removeStylesAndClasses, updateStyling } from "./ApplyStyling";
 import { createCodeblockCodeMirrorExtensions, editingDocumentFold } from "./EditingView";
@@ -35,8 +35,10 @@ export default class CodeStylerPlugin extends Plugin {
 			zoom: document.body.getCssPropertyValue("--zoom-factor"),
 		};
 
-		if (!(await this.app.vault.adapter.exists(EXTERNAL_REFERENCE_PATH))) // Create folder for external references
-			await this.app.vault.adapter.mkdir(EXTERNAL_REFERENCE_PATH);
+		if (!(await this.app.vault.adapter.exists(EXTERNAL_REFERENCE_PATH))) {// Create folder for external references
+			await this.app.vault.adapter.mkdir(this.app.vault.configDir + EXTERNAL_REFERENCE_PATH);
+			await this.app.vault.adapter.write(this.app.vault.configDir + EXTERNAL_REFERENCE_CACHE, JSON.stringify({}));
+		}
 
 		this.executeCodeMutationObserver = executeCodeMutationObserver; // Add execute code mutation observer